create quaiclient test and move filter test to it
